+++ Sources/Swdifft/Marks.swift
@@ -1,0 +1,15 @@
+//
+//  Marks.swift
+//  Swdifft
+//
+//  Created by Yudai Hirose on 2019/07/16.
+//
+
+import Foundation
+
+public var beginLHSMark = "`"
+public var endLHSMark = "`"
+
+public var beginRHSMark = "*"
+public var endRHSMark = "*"
+
